+++ src/types.ts
@@ -1,3 +1,10 @@
+export type NuxtI18nCookieOptions = {
+    maxAge?: number
+    name?: string
+    sameSite?: 'lax' | 'none' | 'strict'
+    secure?: boolean
+}
+
 export type NuxtI18nLocale = {
     code: string
     file: `${string}.ts`
@@ -5,7 +12,10 @@ export type NuxtI18nLocale = {
 }
 
 export interface NuxtI18nOptions {
+    cookie?: NuxtI18nCookieOptions
     defaultLocale?: string
     dir?: string
+    fallbackLocale?: string
     locales?: NuxtI18nLocale[]
+    warnOnMissing?: boolean
 }