+++ test/fixtures/basic/i18n/locales/es.ts
@@ -2,6 +2,7 @@ export default {
     page: {
         home: {
             greeting: 'Hola {{name}}',
+            subtitle: 'Subtitulo ES',
             title: 'Inicio',
         },
     },